+++ app/Http/Controllers/Auth/EmailVerificationPromptController.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace App\Http\Controllers\Auth;
+
+use App\Http\Controllers\Controller;
+use App\Providers\RouteServiceProvider;
+use Illuminate\Http\RedirectResponse;
+use Illuminate\Http\Request;
+use Illuminate\View\View;
+
+class EmailVerificationPromptController extends Controller
+{
+    public function __invoke(Request $request): View|RedirectResponse
+    {
+        if ($request->user()?->hasVerifiedEmail()) {
+            return redirect()->intended(RouteServiceProvider::HOME);
+        }
+
+        return view('auth.verify-email');
+    }
+}