+++ app/Logging/CustomizeFormatter.php
@@ -2,14 +2,26 @@
 
 namespace App\Logging;
 
+use Illuminate\Log\Logger as IlluminateLogger;
 use Monolog\Formatter\JsonFormatter;
 use Monolog\Handler\FormattableHandlerInterface;
 use Monolog\Logger;
 
 class CustomizeFormatter
 {
-    public function __invoke(Logger $logger): void
+    /**
+     * @param  IlluminateLogger|Logger  $logger
+     */
+    public function __invoke($logger): void
     {
+        if ($logger instanceof IlluminateLogger) {
+            $logger = $logger->getLogger();
+        }
+
+        if (! $logger instanceof Logger) {
+            return;
+        }
+
         foreach ($logger->getHandlers() as $handler) {
             if ($handler instanceof FormattableHandlerInterface) {
                 $handler->setFormatter(new JsonFormatter());