+++ config/app.php
@@ -178,6 +178,7 @@ return [
         App\Providers\AppServiceProvider::class,
         App\Providers\AuthServiceProvider::class,
         App\Providers\EventServiceProvider::class,
+        App\Providers\RouteServiceProvider::class,
         App\Providers\Filament\AdminPanelProvider::class,
     ])->toArray(),